Refactor seed script into modular structure
- scripts/seed.ts: 1149 → 273 lines (thin orchestrator)
- scripts/lib/: Reusable utilities
  - supabase.ts: Client setup + env loading
  - imgbb.ts: ImgBB upload utilities
  - storage.ts: Space/tab upload utilities
  - explore-config.ts: Explore page config builder
- scripts/seed-data/communities/: Community DB configs
  - nouns.ts: Nouns config + explore page params
  - clanker.ts: Clanker config
  - example.ts: Template for new communities
- scripts/seed-data/pages/: Static page layouts
- scripts/seed-data/assets/: Asset files for upload

Benefits:
- Easy to add new communities (just add a file)
- Utilities are reusable and testable
- Clear separation of data vs logic
- Self-documenting structure
