+++ scripts/lib/index.ts
@@ -1,0 +1,14 @@
+/**
+ * Seed script utilities index
+ */
+
+export { supabase } from './supabase';
+export { uploadToImgBB, uploadAssets, type AssetMapping } from './imgbb';
+export { 
+  uploadPageConfig, 
+  createNavPageSpace, 
+  getSpaceId,
+  type PageConfigWithTabs,
+} from './storage';
+export { createExplorePageConfig, type TokenInput, type ExplorePageOptions } from './explore-config';
+
