+++ scripts/seed-data/communities/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Community configurations index
+ */
+
+export * from './nouns';
+export * from './clanker';
+export * from './example';
+
