+++ scripts/seed-data/pages/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Page configurations index
+ */
+
+export { nounsHomePage } from './nouns-home';
+export { clankerHomePage } from './clanker-home';
+
